// src/app/models/medicine.ts
export class Medicine {
    // constructor(
    //     private skey: string,
    //     private nombre: string, 
    //     private presentacion: string,
    //     private pastillero: number,
    //     private precio: number,
    //     private stock: number
    //     ) {}
    $key: string;
    nombre: string;
    presentacion: string;
    pastillero: number;
    precio: number;
    stock: number;
}